@@ -1,6 +1,6 @@
 [tool.ruff]
 line-length = 119
-target-version = "py39"
+target-version = "py310"
 lint.ignore = [
     "E501",  # Line too long
     "E731",  # Do not assign a `lambda` expression, use a `def`
