@@ -2,7 +2,7 @@ import logging
 import reprlib
 import time
 from dataclasses import asdict, dataclass, field
-from typing import Any, Literal, Optional
+from typing import Any, Literal
 
 import numpy as np
 import pandas as pd
@@ -136,7 +136,7 @@ class TimeSeriesFeatureGenerator:
         target: str,
         known_covariates_names: list[str],
         float_dtype: str = "float32",
-        num_samples: Optional[int] = 20_000,
+        num_samples: int | None = 20_000,
     ):
         self.target = target
         self.float_dtype = float_dtype
@@ -149,9 +149,9 @@ class TimeSeriesFeatureGenerator:
         self.past_covariates_pipeline = ContinuousAndCategoricalFeatureGenerator()
         # Cat features with cat_count=1 are fine in static_features since they are repeated for all time steps in a TS
         self.static_feature_pipeline = ContinuousAndCategoricalFeatureGenerator(minimum_cat_count=1)
-        self._covariate_metadata: Optional[CovariateMetadata] = None  # type ignore
-        self._train_covariates_real_median: Optional[pd.Series] = None
-        self._train_static_real_median: Optional[pd.Series] = None
+        self._covariate_metadata: CovariateMetadata | None = None  # type ignore
+        self._train_covariates_real_median: pd.Series | None = None
+        self._train_static_real_median: pd.Series | None = None
 
     @property
     def required_column_names(self) -> list[str]:
@@ -279,7 +279,7 @@ class TimeSeriesFeatureGenerator:
             ts_df[column_names] = covariates_real
         return ts_df
 
-    def _impute_static_features(self, static_df: Optional[pd.DataFrame]) -> Optional[pd.DataFrame]:
+    def _impute_static_features(self, static_df: pd.DataFrame | None) -> pd.DataFrame | None:
         """Impute missing values in static features using the median."""
         static_real_names = self.covariate_metadata.static_features_real
         if static_df is not None and static_real_names:
@@ -328,8 +328,8 @@ class TimeSeriesFeatureGenerator:
         return self._impute_covariates(ts_df, column_names=self.covariate_metadata.covariates_real)
 
     def transform_future_known_covariates(
-        self, known_covariates: Optional[TimeSeriesDataFrame]
-    ) -> Optional[TimeSeriesDataFrame]:
+        self, known_covariates: TimeSeriesDataFrame | None
+    ) -> TimeSeriesDataFrame | None:
         assert self._is_fit, f"{self.__class__.__name__} has not been fit yet"
         if len(self.known_covariates_names) > 0:
             assert known_covariates is not None, "known_covariates must be provided at prediction time"
@@ -441,7 +441,7 @@ class PermutationFeatureImportanceTransform(AbstractFeatureImportanceTransform):
         self,
         covariate_metadata: CovariateMetadata,
         prediction_length: int,
-        random_seed: Optional[int] = None,
+        random_seed: int | None = None,
         shuffle_type: Literal["itemwise", "naive"] = "itemwise",
         **kwargs,
     ):
