@@ -1,7 +1,7 @@
 """Common utils and data for all model tests"""
 
 import random
-from typing import Any, Literal, Optional, Union
+from typing import Any, Literal
 
 import numpy as np
 import pandas as pd
@@ -96,7 +96,7 @@ def to_supported_pandas_freq(freq: str) -> str:
 
 
 def get_data_frame_with_item_index(
-    item_list: list[Union[str, int]],
+    item_list: list[str | int],
     data_length: int = 20,
     freq: str = "h",
     start_date: str = "2022-01-01",
@@ -146,8 +146,8 @@ DUMMY_TS_DATAFRAME = mask_entries(get_data_frame_with_item_index(["10", "A", "2"
 
 def get_data_frame_with_variable_lengths(
     item_id_to_length: dict[Any, int],
-    static_features: Optional[pd.DataFrame] = None,
-    covariates_names: Optional[list[str]] = None,
+    static_features: pd.DataFrame | None = None,
+    covariates_names: list[str] | None = None,
     freq: str = "D",
 ):
     tuples = []
@@ -176,10 +176,10 @@ def get_data_frame_with_variable_lengths(
 def get_data_frame_with_covariates(
     item_id_to_length: dict[Any, int] = {1: 10, 5: 20, 2: 30},
     target: str = "target",
-    covariates_cat: Optional[list[str]] = None,
-    covariates_real: Optional[list[str]] = None,
-    static_features_cat: Optional[list[str]] = None,
-    static_features_real: Optional[list[str]] = None,
+    covariates_cat: list[str] | None = None,
+    covariates_real: list[str] | None = None,
+    static_features_cat: list[str] | None = None,
+    static_features_real: list[str] | None = None,
 ):
     data = get_data_frame_with_variable_lengths(item_id_to_length)
     data.rename(columns={"target": target}, inplace=True)
@@ -205,7 +205,7 @@ ITEM_ID_TO_LENGTH = {"D": 22, "A": 50, "C": 10, "B": 17}
 DUMMY_VARIABLE_LENGTH_TS_DATAFRAME = mask_entries(get_data_frame_with_variable_lengths(ITEM_ID_TO_LENGTH))
 
 
-def get_static_features(item_ids: list[Union[str, int]], feature_names: list[str]):
+def get_static_features(item_ids: list[str | int], feature_names: list[str]):
     features = {}
     for idx, feat_name in enumerate(feature_names):
         if idx % 2 == 0:
